fixed an integer overflow bug when writing large files > 2147483648 bytes, which causes an exception to be thrown by the caller

+++ src/main/java/com/ning/http/client/multipart/FilePart.java
@@ -101,7 +101,7 @@ public class FilePart extends AbstractFilePart {
 
         handler.start();
 
-        int length = 0;
+        long length = 0;
 
         length += MultipartUtils.writeBytesToChannel(target, generateFileStart(boundary));
